refactor: extract disk-layout into shared package
- Move all disk-layout helpers (hourIdToPath, pathToHourId,
  enumerateRange, latestOnDisk, cleanPartials) from
  data-pipeline/fetcher/lib/disk-layout.js to
  data-pipeline/shared/disk-layout/index.js
- Add data-pipeline/shared/disk-layout/package.json (name: disk-layout)
- Update fetcher.js and lib/download.js to import from the new location
  via relative paths (no workspace config needed)
- Delete the old fetcher-local disk-layout.js

No behaviour changes; all existing smoke tests remain valid.

Closes #196
