+++ data-pipeline/shared/disk-layout/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "disk-layout",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "index.js"
+}
